+++ tests/testthat/test-flatten.R
@@ -1,16 +1,3 @@
-test_that("flatten functions are deprecated", {
-  expect_snapshot({
-    . <- flatten(list())
-    . <- flatten_lgl(list())
-    . <- flatten_int(list())
-    . <- flatten_dbl(list())
-    . <- flatten_chr(list())
-    . <- flatten_dfr(list())
-    . <- flatten_dfc(list())
-    . <- flatten_df(list())
-  })
-})
-
 test_that("input must be a list", {
   local_options(lifecycle_verbosity = "quiet")
 
@@ -26,15 +13,11 @@ test_that("contents of list must be supported types", {
 })
 
 test_that("each second level element becomes first level element", {
-  local_options(lifecycle_verbosity = "quiet")
-
   expect_equal(flatten(list(1:2)), list(1, 2))
   expect_equal(flatten(list(1, 2)), list(1, 2))
 })
 
 test_that("can flatten all atomic vectors", {
-  local_options(lifecycle_verbosity = "quiet")
-
   expect_equal(flatten(list(F)), list(F))
   expect_equal(flatten(list(1L)), list(1L))
   expect_equal(flatten(list(1)), list(1))
@@ -44,23 +27,17 @@ test_that("can flatten all atomic vectors", {
 })
 
 test_that("NULLs are silently dropped", {
-  local_options(lifecycle_verbosity = "quiet")
-
   expect_equal(flatten(list(NULL, NULL)), list())
   expect_equal(flatten(list(NULL, 1)), list(1))
   expect_equal(flatten(list(1, NULL)), list(1))
 })
 
 test_that("names are preserved", {
-  local_options(lifecycle_verbosity = "quiet")
-
   expect_equal(flatten(list(list(x = 1), list(y = 1))), list(x = 1, y = 1))
   expect_equal(flatten(list(list(a = 1, b = 2), 3)), list(a = 1, b = 2, 3))
 })
 
 test_that("names of 'scalar' elements are preserved", {
-  local_options(lifecycle_verbosity = "quiet")
-
   out <- flatten(list(a = list(1), b = list(2)))
   expect_equal(out, list(a = 1, b = 2))
 
@@ -72,8 +49,6 @@ test_that("names of 'scalar' elements are preserved", {
 })
 
 test_that("child names beat parent names", {
-  local_options(lifecycle_verbosity = "quiet")
-
   out <- flatten(list(a = list(x = 1), b = list(y = 2)))
   expect_equal(out, list(x = 1, y = 2))
 })
@@ -88,8 +63,6 @@ test_that("must be a list", {
 })
 
 test_that("can flatten all atomic vectors", {
-  local_options(lifecycle_verbosity = "quiet")
-
   expect_equal(flatten_lgl(list(F)), F)
   expect_equal(flatten_int(list(1L)), 1L)
   expect_equal(flatten_dbl(list(1)), 1)
@@ -97,8 +70,6 @@ test_that("can flatten all atomic vectors", {
 })
 
 test_that("preserves inner names", {
-  local_options(lifecycle_verbosity = "quiet")
-
   expect_equal(
     flatten_dbl(list(c(a = 1), c(b = 2))),
     c(a = 1, b = 2)
@@ -109,7 +80,6 @@ test_that("preserves inner names", {
 # data frame flatten ------------------------------------------------------
 
 test_that("can flatten to a data frame with named lists", {
-  local_options(lifecycle_verbosity = "quiet")
   skip_if_not_installed("dplyr")
 
   dfs <- list(c(a = 1), c(b = 2))